Added causal graph plot in the notebook and moved some code in a utils.py file
